Change the way dependency injection works in presenter
Move away from weird `receiveDependencies` method to the using the
constructor. The `receiveDependencies` methods keep working as a
fallback.

The constructor does not receive the initial collection anymore, and the
collectFromCollection is no longer called. This was bad practice from
the beginning and should not have been used in the first place. With the
"present" helper methods it is possible to do everything that
`collectFromCollection` did.