Stop depending on `@depends` in tests to insert records
There was a single test responsible to create the records in the
database for each test class, this introduces problem when tests
are reliant on eachother to have the right numbers/IDs. The new
setup creates a completely fresh database instance, and insert the
needed records everytime. A helper method is created to simplify
this.