Rework query conditions
- Use `Condition`s for all conditions added to a query, the string
  approach works by converting them to a `Condition`
- Allow passing `Condition`s to the several methods that accept a
  condition in some way, like `where` and `innerJoin`.

Using the `Condition`s directly allows allows for more complicated
conditions that were not possible with the array approach, like nested
AND/OR clauses with multiple values.

The generated queries are slightly different in some cases, the new
approach allows for better separation/detection of multiple conditions.
This means that a lot of the redundant parentheses are now no longer
included, they are only added around the `Multiple` condition to make
sure they stay grouped together.