Always wrap raw query conditions in parentheses
Raw query conditions can contain anything including stuff that would
break the specificity rules of AND/OR. If a raw conditions has an OR
clause inside of it, then it would be possible that the OR applies to
something else when combined with another AND clauses.

For example, the following changes based on whether there are
parentheses around the raw condition:

```php
$query->where([
    '1 OR 2',
    '3',
]);
```

Would produce: `1 OR 2 AND 3`, which very likely it not the intended
result. By always adding the parentheses the ambiguity is removed:
`(1 OR 2) AND (3)`.

This only needs to happen with raw conditions, the other conditions are
generated and already have parentheses when needed.