Include clauses in presenter queries
Previously the presenters would only use the clauses to filter and/or
order the found entities after the query was run. This might be wasteful
if the clause is used to only select a small subset of the query result.
Using the clause in the query makes the query less wasteful, this
behavior is skipped if the entity pool has a request for different
clauses for the same entities.